# huff0: Speed up compress1xDo